Add README and dependencies to S-Comp project
Introduced a README.md file providing an overview and specifics of the S-Comp project, a simple compiler in Scala. Additionally, added library dependencies 'scalactic' and 'scalatest' to build.sbt. The README serves as an introductory guide and manual, while the new dependencies provide testing capabilities for the development of the S-Comp project.