Add boolean expressions handling and respective tests
Introduced a new file, Bool.scala which adds handling for boolean expressions in the expression language. This includes objects that can represent `True` and `False` values, along with their prefix representation. Accompanying tests were added in BoolTest.scala to ensure correct implementation of these objects. This provides a new way to deal with boolean values distinctly in the system.