Implement 'ret' Instruction and refactor compilation logic
Added the 'ret' instruction representation in the `Instruction` sealed trait, and utilised it in the assembly generation process. Additionally, restructured the `compileExpression` function for better readability and made it only accessible within its package. This improves the modular separation of responsibility in the codebase. Also, minor changes in arrangement of import statements were made.