✨🔨📝 Implement multiplication operation in AST and compiler
The multiplication operation has been added in the abstract syntax tree (AST) and the compiler. This includes creating a new file for the 'TimesImpl' trait, and adding a 'Times' case class to 'Expression.scala'. The pattern matching in 'compileExpression' has also been updated to account for the multiplication operation. This update enhances the variety of expressions that can be represented and processed.